requirements.txt
Flask
pandas
textblob
Flask-MySQLdb
python-dotenv
waitress